backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
email-validator
PyJWT[crypto]
pytest
# MongoDB ODM/driver
beanie
motor
python-dotenv
regex
passlib[bcrypt]
strawberry-graphql[fastapi]
httpx
python-jose[cryptography]
